scraper/requirements.txt
# iyisiniye Scraper Bagimliliklari
# pip install -r requirements.txt

# Web Scraping
scrapy>=2.11.0
scrapy-playwright>=0.0.42
playwright>=1.49.0
httpx>=0.28.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
itemadapter>=0.9.0

# NLP & ML
transformers>=4.47.0
torch>=2.5.0
rapidfuzz>=3.10.0

# Veri Isleme
pandas>=2.2.0
pydantic>=2.10.0

# Veritabani
psycopg2-binary>=2.9.0

# Altyapi
python-dotenv>=1.0.0
loguru>=0.7.0
tenacity>=9.0.0

# Gelistirme
pytest>=8.3.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
mypy>=1.13.0
